Split out non web parts into node/externalTerminal.contribution
