+++ src/vs/workbench/contrib/externalTerminal/node/externalTerminal.ts
@@ -1,0 +1,6 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export const DEFAULT_TERMINAL_OSX = 'Terminal.app';
